Address review feedback: generalize foreign Clang type check, add explanation comment, and fold tests
